Start version control package mppAssTest
This package will be used to keep in an acceptable format the functionalities developped for the third chapter.

I started by doing some function to produce the plink .bed file, to compute the LDAK weights, to compute manually the kinship matrix and finally to compute the single SNP association scan with extra functions to plot the results.