Commit bfre modifying the haplotype determination fct
I review the haplotype determination function (haplo_block). I stop to use SelectionTools because it has some problem and I do not understand how it function.

I start a new function (haplo_block2) which allow to make haplotype partition using mk number or cM distance with adjacent blocks or sliding window.

For both the haplotype model and the kernel association test, the idea is to produce some partition files that will be used after to compute the models.